test-generator-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
supabase==2.9.0
google-genai==0.8.0
python-dotenv==1.0.1
python-multipart==0.0.9
httpx==0.27.2
sentence-transformers==3.0.1